Update jwcrypto dependency to version 1.5.0 (#105)
This change updates the jwcrypto dependency in pyproject.toml from a patched version to the official release version 1.5.0. The previous patched version was used to support the brainpool curves, but now the official release includes this support.

This update has implications for any code that relies on jwcrypto for JSON Web Token (JWT) cryptography and signing operations. Developers can now use the official release of jwcrypto instead of relying on a patched version.

See the `jwcrypto^1.5.0` changelog for reference: https://github.com/latchset/jwcrypto/releases/tag/v1.5.0

Additionally, this change ensures compatibility with other dependencies and allows for easier maintenance and future updates of the project.
Quick note on the `poetry.lock`: Python-poetry 1.6.1 now takes take of category matching, this is causing the discontinuation of the `category` property.

Co-authored-by: Lukas Wolfsteiner <lukas.wolfsteiner@awesome-technologies.de>